client/react: resolve "can resolve ticket?" on fetch with a batch request
It's a bit odd at the moment, since we're not using any of the tickets'
properties to make the decision -- so we might as well just have one eval.
But it openes up our policies to be more fine grained...and it is the only
use case of batch eval in tickethub's UI that I have found so far.

Signed-off-by: Stephan Renatus <stephan@styra.com>