client/react: add batch demo page
It's not exposed through the UI, but you can do go either
- http://127.0.0.1:3000/demo or
- http://127.0.0.1:3000/demo?batch=true

for the non-batching/batching load.

When playing with it, it can be useful to throttle network requests:
otherwise, there's little difference, I think.

Also, I don' yet understand why it's updating in the way it is updating,
so slowly. I'll think about it more. Maybe there are performance issues
to look into.

Signed-off-by: Stephan Renatus <stephan@styra.com>