server/node: use ucast conditions (#383)
This isn't completely done -- we'll want to add more of this, beyond a
single list endpoint getting filtered. But it's a start, it works, and it has
tests, so let's get it in.

The policies/tickets.rego file is a bit messy because I've left other conditions
examples in, but we'll probably revisit that file soon anyways.

* use ucastToPrisma() from `@styra/ucast-prisma`
* server/node: use node-22 for docker image
* tests: add conditions-exercising tests
* clients: "fix" ceasar spelling

It's still wrong, but now it's consistent. Before, the resolver-role
user "ceasar" wasn't usable from the UI code.

Signed-off-by: Stephan Renatus <stephan@styra.com>