server/springboot: alter "list tickets" SQL query on the fly
One learning is that we cannot rely on the "method name to SQL query"
magic that JPA supports -- rewriting queries using QueryRewriter only
works for JPQL or "native" (=dialect of the database) queries.

lmgty:
JPQL (Java Persistence Query Language) is a query language used to perform
database operations on entities defined in JPA (Java Persistence API). It
is similar to SQL but operates on the entity objects instead of the
database tables.

Signed-off-by: Stephan Renatus <stephan@styra.com>