tests/api: make mask assertion a bit looser
One test suite creates a ticket, the other asserts something on the
number of tickets returned. That's inherently racy, and makes the
latter tests fail unless they're run in parallel, and we're also lucky.

Now, the count-asserting tests (masks.hurl) only check the first 5
tickets. Should there be 6 because the tests run in parallel, the
assertion still holds true for the first 5. (It would also hold true for
the sixth ticket, but that isn't asserted.)

We're also removing the `all_count` captures, since they're not used.

Signed-off-by: Stephan Renatus <stephan@styra.com>